+++ Source/WildShift/DefOfs.cs
@@ -10,7 +10,6 @@ namespace WildShift
         public static HediffDef WildShift_Transformed;
         public static HediffDef WildShift_FieldShapeshifterMarker;
         public static AbilityDef WildShift_Shift;
-        public static IncidentDef WildShift_ShapeshifterJoin;
         public static ThinkTreeDef WildShift_TransformedDrafted;
 
         static WildShiftDefOf()